+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Telemetry and car status packages don't raise when no lap is set
 - Added unknown game mode 100
 - Round setup float values
+- Fixed lap sector 3 recompute type issue
 
 
 ## 3.0.2 - 2022-07-01